completed user authorization on tasks

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,5 @@
 import { Repository } from 'typeorm';
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { BadRequestException } from '@nestjs/common/exceptions';
 
@@ -8,6 +8,8 @@ import { TaskStatus } from './task-status.enum';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 
+import User from '../auth/user.entity';
+
 @Injectable()
 export class TasksService {
   constructor(
@@ -15,9 +17,11 @@ export class TasksService {
     private taskRepository: Repository<Task>,
   ) {}
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
     const { status, search } = filterDto;
     const query = this.taskRepository.createQueryBuilder('task');
+
+    query.where('task.userId = :userId', { userId: user.id });
 
     if (status) {
       query.andWhere('task.status = :status', { status });
@@ -34,33 +38,40 @@ export class TasksService {
     return tasks;
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOneBy({ id });
-    if (!found) throw new BadRequestException(`Task with ID '${id}' not found`);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOneBy({ id, userId: user.id });
+    if (!found) throw new NotFoundException(`Task with ID '${id}' not found`);
 
     return found;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
     const task = new Task();
     task.title = createTaskDto.title;
     task.description = createTaskDto.description;
     task.status = TaskStatus.OPEN;
+    task.user = user;
     await task.save();
     // this.taskRepository.create(task);
     // await this.taskRepository.save(task);
+
+    delete task.user;
     return task;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0)
       throw new BadRequestException(`Task with ID '${id}' not found`);
